Improve component composition
