+++ src/pages/ChainStats/components/StatsChart/index.ts
@@ -1,0 +1,2 @@
+export * from './StatsChart';
+export { default } from './StatsChart';
