+++ src/pages/ChainStats/components/index.ts
@@ -1,3 +1,4 @@
 export * from './ChartCoin';
 export * from './ChartDurationPicker';
 export * from './ChartMetrics';
+export * from './StatsChart';
