+++ src/pages/ChainStats/constants.ts
@@ -1,0 +1,22 @@
+export const DURATION_OPTIONS = [
+  {
+    label: '1D',
+    value: '1d',
+  },
+  {
+    label: '1W',
+    value: '1w',
+  },
+  {
+    label: '1M',
+    value: '1m',
+  },
+  {
+    label: '1Y',
+    value: '1y',
+  },
+  {
+    label: 'ALL',
+    value: 'all',
+  },
+];
