+++ src/pages/ChainStats/hooks/useNetworkStatsChartData/index.ts
@@ -1,0 +1,2 @@
+export * from './useNetworkStatsChartData';
+export { default } from './useNetworkStatsChartData';
